+++ src/derivatives_trading_coin_futures/rest_api/models/all_orders_response_inner.rs
@@ -53,7 +53,7 @@ pub struct AllOrdersResponseInner {
     /// Position side
     #[serde(rename = "positionSide", skip_serializing_if = "Option::is_none")]
     pub position_side: Option<String>,
-    /// Enum：completed，processing
+    /// Status.
     #[serde(rename = "status", skip_serializing_if = "Option::is_none")]
     pub status: Option<String>,
     /// please ignore when order type is `TRAILING_STOP_MARKET`