+++ src/derivatives_trading_coin_futures/rest_api/models/cancel_order_response.rs
@@ -44,7 +44,7 @@ pub struct CancelOrderResponse {
     /// Position side
     #[serde(rename = "positionSide", skip_serializing_if = "Option::is_none")]
     pub position_side: Option<String>,
-    /// Enum：completed，processing
+    /// Status.
     #[serde(rename = "status", skip_serializing_if = "Option::is_none")]
     pub status: Option<String>,
     /// please ignore when order type is `TRAILING_STOP_MARKET`