+++ src/derivatives_trading_coin_futures/rest_api/models/users_force_orders_response_inner.rs
@@ -26,7 +26,7 @@ pub struct UsersForceOrdersResponseInner {
     /// Pair
     #[serde(rename = "pair", skip_serializing_if = "Option::is_none")]
     pub pair: Option<String>,
-    /// Enum：completed，processing
+    /// Status.
     #[serde(rename = "status", skip_serializing_if = "Option::is_none")]
     pub status: Option<String>,
     /// Client order ID.