+++ src/derivatives_trading_coin_futures/websocket_api/models/cancel_order_response.rs
@@ -21,7 +21,6 @@ pub struct CancelOrderResponse {
     /// positionId
     #[serde(rename = "id", skip_serializing_if = "Option::is_none")]
     pub id: Option<String>,
-    /// Enum：completed，processing
     #[serde(rename = "status", skip_serializing_if = "Option::is_none")]
     pub status: Option<i64>,
     #[serde(rename = "result", skip_serializing_if = "Option::is_none")]