+++ src/derivatives_trading_coin_futures/websocket_api/models/modify_order_response_result.rs
@@ -28,7 +28,7 @@ pub struct ModifyOrderResponseResult {
     /// Pair
     #[serde(rename = "pair", skip_serializing_if = "Option::is_none")]
     pub pair: Option<String>,
-    /// Enum：completed，processing
+    /// Status.
     #[serde(rename = "status", skip_serializing_if = "Option::is_none")]
     pub status: Option<String>,
     /// Client order ID.