+++ src/derivatives_trading_portfolio_margin/rest_api/models/modify_cm_order_response.rs
@@ -26,7 +26,7 @@ pub struct ModifyCmOrderResponse {
     /// Pair.
     #[serde(rename = "pair", skip_serializing_if = "Option::is_none")]
     pub pair: Option<String>,
-    /// Enum：completed，processing
+    /// Status.
     #[serde(rename = "status", skip_serializing_if = "Option::is_none")]
     pub status: Option<String>,
     /// Client Order ID.