+++ src/derivatives_trading_portfolio_margin/rest_api/models/new_margin_order_response.rs
@@ -44,7 +44,7 @@ pub struct NewMarginOrderResponse {
         skip_serializing_if = "Option::is_none"
     )]
     pub cummulative_quote_qty: Option<String>,
-    /// Enum：completed，processing
+    /// Status.
     #[serde(rename = "status", skip_serializing_if = "Option::is_none")]
     pub status: Option<String>,
     /// Time In Force.