+++ src/derivatives_trading_portfolio_margin/rest_api/models/new_um_order_response.rs
@@ -44,7 +44,7 @@ pub struct NewUmOrderResponse {
     /// BOTH means that it is the position of One-way Mode
     #[serde(rename = "positionSide", skip_serializing_if = "Option::is_none")]
     pub position_side: Option<String>,
-    /// Enum：completed，processing
+    /// Status.
     #[serde(rename = "status", skip_serializing_if = "Option::is_none")]
     pub status: Option<String>,
     /// Trade symbol, if existing.