+++ src/derivatives_trading_portfolio_margin/websocket_streams/models/algo_order_update_ao.rs
@@ -94,6 +94,9 @@ pub struct AlgoOrderUpdateAo {
     /// Algo order failed reason
     #[serde(rename = "rm", skip_serializing_if = "Option::is_none")]
     pub rm: Option<String>,
+    /// Is activated or not, only meaningful for trailing order
+    #[serde(rename = "ia", skip_serializing_if = "Option::is_none")]
+    pub ia: Option<bool>,
 }
 
 impl AlgoOrderUpdateAo {
@@ -126,6 +129,7 @@ impl AlgoOrderUpdateAo {
             tt: None,
             gtd: None,
             rm: None,
+            ia: None,
         }
     }
 }